IO of indexMetaFile

+++ directory/mmapDirectory.go
@@ -6,67 +6,73 @@ import (
 	"os"
 )
 
+const InitMapSize = 4 * 1024
+
 type MMapDirectory struct {
 	ctx  *Context
-	core support.MMapCore
+	core map[string]*support.MMapCore
 }
 
-func NewMMapDirectory(ctx *Context) Directory {
-	return &MMapDirectory{ctx: ctx}
+func (d *MMapDirectory) Reader(file *FileSlice) DataIO {
+	return d.core[file.Path]
 }
 
-func (d *MMapDirectory) Read(path string) DataIO {
-	return d.doMap(path)
+func (d *MMapDirectory) Writer(file *FileSlice) DataIO {
+	return d.core[file.Path]
 }
 
-func (d *MMapDirectory) Write(path string) DataIO {
-	return d.doMap(path)
+func (d *MMapDirectory) Delete(file *FileSlice) {
+	delete(d.core, file.Path)
 }
 
-func (d *MMapDirectory) Delete(path string) {
-
-}
-
-func (d *MMapDirectory) FileOp(path string) *FileSlice {
-	core := d.doMap(path)
-	slice := &FileSlice{File: core.File, Path: path, Io: core}
+func (d *MMapDirectory) FileOp(path string, cap int) *FileSlice {
+	core := d.doMap(path, cap)
+	slice := &FileSlice{File: core.File, Path: path}
+	d.core[slice.Path] = core
 	return slice
 }
 
-func (d *MMapDirectory) Exists(path string) (bool, error) {
-	f, _ := d.openFile(path)
+func (d *MMapDirectory) Exists(path string, create bool) (bool, error) {
+	f, _ := d.openFile(path, create)
 	if f == nil {
 		return false, errors.New("file not found" + path)
 	}
-	return true, nil
+	return f != nil, nil
 }
 
-func (d *MMapDirectory) doMap(path string) *support.MMapCore {
-	file, _ := d.openFile(path)
-	m, err := d.core.Issued(file)
+func (d *MMapDirectory) doMap(path string, cap int) *support.MMapCore {
+	file, _ := d.openFile(path, false)
+	if cap < 0 {
+		info, _ := file.Stat()
+		cap = int(info.Size())
+	}
+	m, err := support.IssuedMMap(file, cap)
 	if err != nil {
 		d.ctx.Log.Fatalln("cannot do mmap", err.Error())
 	}
 	return m
 }
 
-func (d *MMapDirectory) openFile(path string) (*os.File, os.FileInfo) {
-	f, err := os.OpenFile(path, os.O_RDWR|os.O_CREATE, 0777)
+func (d *MMapDirectory) openFile(path string, create bool) (*os.File, os.FileInfo) {
+	flag := os.O_RDWR
+	if create {
+		flag = os.O_RDWR | os.O_CREATE
+	}
+	f, err := os.OpenFile(path, flag, 0777)
 	if err != nil {
-		file := d.maybeCreateFile(err, path)
+		file := d.maybeCreateFile(err, path, create)
 		if file != nil {
 			info, _ := file.Stat()
 			return file, info
 		}
-		d.ctx.Log.Fatalln("open file fail,", err.Error())
 		return nil, nil
 	}
 	info, _ := f.Stat()
 	return f, info
 }
 
-func (d *MMapDirectory) maybeCreateFile(err error, path string) *os.File {
-	if os.IsNotExist(err) {
+func (d *MMapDirectory) maybeCreateFile(err error, path string, create bool) *os.File {
+	if os.IsNotExist(err) && create {
 		file, err := os.Create(path)
 		if err != nil {
 			panic(err.Error())
